VERY IMPORTANT - Implement Ride Matching + Display
VERY IMPORTANT COMMIT - In server.js, add handling for api/ride/search/ path for matching inputted fields with users in the database. Inputted fields are then stored into the user's spot in the database. In RideRequestsPage.js, when a form is submitted, the matching users are mapped to the page. TravelSuggestionsPage.js, fixed warnings from useEffect.