MessageCenterPage backend and frontend rework
For message center, the current users chatLogs are outputted to the sidebar in the page. The user clicking a chatlog will display that conversations messages. Currently no messages in database + NOTE: sending a memssage will ERROR because the logic in the front+backend is not implemented yet. This is a WIP, likely to be finished in the next commit.